# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Content Generation Service",
    "description": "AI-powered content generation service",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://content-generation.mcpfactory.org"
    }
  ],
  "security": [
    {
      "ApiKeyAuth": []
    }
  ],
  "components": {
    "schemas": {
...
      "GenerateRequest": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "description": "Which stored prompt to use, e.g. 'cold-email' or 'welcome-email'"
          },
          "variables": {
            "type": "object",
            "additionalProperties": {
              "nullable": true
            },
            "description": "Variable values to substitute into the prompt template. Any JSON values allowed — strings, arrays, or objects. Caller decides the shape per variable. Objects and arrays are rendered as readable markdown into the prompt; the LLM reads whatever's provided. Multibrand is the default in this platform, so brand-related variables typically receive arrays or objects, not scalars. Per-template input expectations are published via GET /platform-prompts?type=<type> (.variables: Array<{ name, description }>). When values are string-typed, recognised keys may also populate dedicated dashboard columns: leadFirstName, leadLastName, leadTitle, leadCompanyName, leadCompanyIndustry, organizationDomain, clientCompanyName."
          },
          "brandIds": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Brand UUIDs associated with this generation. Falls back to parsed x-brand-id header (CSV) if omitted."
          },
          "campaignId": {
            "type": "string"
          },
          "apolloEnrichmentId": {
            "type": "string"
          },
          "leadId": {
            "type": "string",
            "description": "Lead-service correlation ID, unique per campaign"
          },
          "idempotencyKey": {
            "type": "string"
          },
          "workflowSlug": {
            "type": "string"
          },
          "featureSlug": {
            "type": "string",
            "description": "Feature slug for tracking"
          }
        },
        "required": [
          "type",
          "variables"
        ]
      },
...
    },
...
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Api-Key"
      }
    }
  },
  "paths": {
...
    "/generate": {
      "post": {
        "tags": [
          "Content Generation"
        ],
        "summary": "Generate content using a stored prompt template with variable substitution",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "x-org-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "x-user-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "x-run-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "x-campaign-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Comma-separated brand UUIDs (e.g. 'uuid1,uuid2')"
            },
            "required": false,
            "description": "Comma-separated brand UUIDs (e.g. 'uuid1,uuid2')",
            "name": "x-brand-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "x-workflow-slug",
            "in": "header"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "x-feature-slug",
            "in": "header"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Generated content",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenerateResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "402": {
            "description": "Insufficient credits (platform costSource only)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InsufficientCreditsResponse"
                }
              }
            }
          },
          "404": {
            "description": "Prompt not found for this type",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/generate-expert-quote-pitch": {
      "post": {
        "tags": [
          "Content Generation"
        ],
        "summary": "Generate a journalist-quote pitch (Featured.com 100-2500 char constraint)",
        "description": "Renders the stored 'expert-quote-pitch' prompt with brand + request inputs, sends to chat-service for free-text generation, and enforces a 100-2500 character output range. If the first attempt is out of range, retries once with a corrective nudge. Returns 400 ExpertQuotePitchLengthErrorResponse if both attempts are out of range — no truncation that breaks meaning.",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "x-org-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "x-user-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "x-run-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "x-campaign-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "x-brand-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "x-workflow-slug",
            "in": "header"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "x-feature-slug",
            "in": "header"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateExpertQuotePitchRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Generated pitch within the configured char range",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenerateExpertQuotePitchResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request body or pitch length out of range after retry",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExpertQuotePitchLengthErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "402": {
            "description": "Insufficient credits",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InsufficientCreditsResponse"
                }
              }
            }
          },
          "404": {
            "description": "expert-quote-pitch template not registered",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/generations": {
      "get": {
        "tags": [
          "Content Generation"
        ],
        "summary": "List generations with filters",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "runId",
            "in": "query"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "campaignId",
            "in": "query"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "brandId",
            "in": "query"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "x-org-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "x-user-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "x-run-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "x-campaign-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Comma-separated brand UUIDs (e.g. 'uuid1,uuid2')"
            },
            "required": false,
            "description": "Comma-separated brand UUIDs (e.g. 'uuid1,uuid2')",
            "name": "x-brand-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "x-workflow-slug",
            "in": "header"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "x-feature-slug",
            "in": "header"
          }
        ],
        "responses": {
          "200": {
            "description": "List of generations",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenerationsListResponse"
                }
              }
            }
          },
          "400": {
            "description": "At least one filter required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  }
}